% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddCensusDemographics.R
\name{AddCensusDemographics}
\alias{AddCensusDemographics}
\title{Add Census demographics data to your simple features object}
\usage{
AddCensusDemographics(map, year = 2010, state, county = NULL,
  demographicvars = NULL, api_key = NULL)
}
\arguments{
\item{map}{A simple features map covering a part of one (and only one) state.}

\item{year}{The Census year from which you want data.}

\item{state}{The state that contains map. Accepts FIPS code and two letter abreviation.}

\item{county}{The FIPS code of the county/counties containing map. This defaults to NULL, but the function runs more quickly the fewer counties downloaded.}

\item{demographicvars}{The demographic variables you want to add. If NULL (the default) it downloads the total population which is "P0010001" for 2010 and "P001001" for 1990 and 2000. Codes can be looked up at https://api.census.gov/data.html by clicking on "variables" next to the Census year you want.}

\item{api_key}{Your census API key. Works with NULL (default), but you are subject to limits without a key.}
}
\value{
A new simple features object with U.S. Census demographic data added.
}
\description{
Add Census demographics data to your simple features object
}
