\name{AddCensusDemographics}
\alias{AddCensusDemographics}
\title{AddCensusDemographics}
\usage{

MapWithDemographics <- AddCensusDemographics(map = map,state = state,county = county)


}
\description{
Returns a map with Census demographics added.
Go to: https://api.census.gov/data.html and click variables next to the dataset you will use to see available demographic variables.

}
\examples{
state="MA"
county=17
state
county

}
